requirements.txt
pyyaml
python-dotenv
google-genai
pydantic
plotly
streamlit